Add theme handler class
