+++ src/generators/tkinter_generator.py
@@ -8,6 +8,8 @@ import tkinter as tk
 from tkinter import ttk
 from tkinter import filedialog, messagebox
 from collections import deque
+
+from generators.theme_handler import ThemeHandler
 from utils import util_functions, globals as my_globals
 from floors.floor import Floor
 from floors.tkinter_floor import TkinterFloor
@@ -50,47 +52,8 @@ class TkinterGenerator(Generator):
         self._next_seed.set(secrets.token_hex(16))
         self._name.set(self._output_file_name)
         self._menu_bar = tk.Menu(self._tk)
-        self._current_theme = "light"
-        self.apply_theme()
+        self._theme_handler = ThemeHandler(self._tk, self._canvas)
         self._tk.protocol('WM_DELETE_WINDOW', self.quit)
-
-    def apply_theme(self) -> None:
-        """
-        Applies the current theme to the application.
-        """
-        if self._current_theme == "light":
-            self._set_light_theme()
-        else:
-            self._set_dark_theme()
-
-    def _set_light_theme(self) -> None:
-        self._tk.configure(background="light gray", highlightbackground="black")
-        style = ttk.Style(self._tk)
-        style.configure(
-            "TButton", background="white", foreground="black", hover="black"
-        )
-        style.configure("TFrame", background="light gray", foreground="black")
-        style.theme_use("default")
-        self._canvas.configure(background="white", highlightbackground="black")
-
-    def _set_dark_theme(self) -> None:
-        self._tk.configure(background="dark gray", highlightbackground="white")
-        self._canvas.configure(background="black", highlightbackground="white")
-        style = ttk.Style(self._tk)
-        style.configure("TFrame", background="dark gray", foreground="white")
-        style.configure("TButton", background="black", foreground="white")
-        # set color on hover
-        style.map("TButton", background=[("active", "gray")])
-
-    def switch_theme(self) -> None:
-        """
-        Switches the theme of the application.
-        """
-        if self._current_theme == "light":
-            self._current_theme = "dark"
-        else:
-            self._current_theme = "light"
-        self.apply_theme()
 
     def _create_floor(self) -> None:
         """
@@ -223,7 +186,6 @@ class TkinterGenerator(Generator):
         self.add_information_frame()
         self.add_buttons()
         self.create_menu_bar()
-        self.apply_theme()
         self._tk.mainloop()
 
     def quit(self) -> None:
@@ -303,6 +265,6 @@ class TkinterGenerator(Generator):
             label="Save As", command=self.save, accelerator="Ctrl+Shift+s"
         )
         menu_bar.add_command(
-            label="Switch Theme", command=self.switch_theme, accelerator="Ctrl+t"
+            label="Switch Theme", command=self._theme_handler.switch_theme, accelerator="Ctrl+t"
         )
         menu_bar.add_command(label="Exit", command=self.quit, accelerator="Ctrl+q")
